[HAL] Move HALs to src
