+++ src/hal/Arduino/ArduinoHal.cpp
