+++ src/hal/Arduino/ArduinoHal.h
