+++ src/modules/SX126x/STM32WLx_Module.cpp
@@ -9,7 +9,7 @@ This file is licensed under the MIT License: https://opensource.org/licenses/MIT
 
 #if !RADIOLIB_EXCLUDE_STM32WLX
 
-#include "../../ArduinoHal.h"
+#include "hal/Arduino/ArduinoHal.h"
 
 // This defines some dummy pin numbers (starting at NUM_DIGITAL_PINS to
 // guarantee these are not valid regular pin numbers) that can be passed
